fix: ajusta validação na rota de deleção

+++ app.py
@@ -92,6 +92,8 @@ def update_user(id_user):
 @login_required
 def delete_user(id_user):
     user = User.query.get(id_user)
+    if id_user == current_user.id:
+        return jsonify({"message": "Deleção não permitida"}), 403
     if user and id_user != current_user.id:
         db.session.delete(user)
         db.session.commit()
